validators and help page

+++ app/Http/Controllers/User/IssueController.php
@@ -16,6 +16,8 @@ use App\User;
 use Carbon\Carbon;
 use DB;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Validator;
+
 
 class IssueController extends Controller
 {
@@ -92,6 +94,23 @@ class IssueController extends Controller
     public function create(Request $request)
     {
         $issue = new Issue();
+
+        $rules = [  
+            'title' => 'required|string',
+            'author' => 'required|integer|exists:user,id',
+            'description' => 'string|nullable',
+            'list' => 'required|integer|exists:issue_list,id'
+        ];
+
+        $messages = [
+            'required' => ":attribute is required for the issue. Please choose it"
+        ];
+
+        $validator = Validator::make($request->all(), $rules, $messages);
+            if ($validator->fails()) {
+            return response()->json(['errors' => $validator->errors(), 404]);
+        }
+            
         $issue->name = $request->input('title');
         $issue->author_id = $request->input('author');
         $issue->description = $request->input('description');
@@ -103,6 +122,22 @@ class IssueController extends Controller
 
     public function addList(Request $request, $id)
     {
+        
+        $rules = [  
+            'name' => 'required|string|min:1',
+        ];
+
+        $messages = [
+            'name.required' => "Please choose a name for the issue list",
+        ];
+        $validator = Validator::make($request->all(), $rules, $messages);
+
+        if ($validator->fails()) {
+            return response()->json(['errors' => $validator->errors(), 404]);
+        }
+        return response()->json([$request->all()]);
+
+        
         $project = Project::findOrFail($id);
 
         $issueList = new IssueList();
@@ -126,6 +161,27 @@ class IssueController extends Controller
     {
         $issue = Issue::findOrFail($id);
 
+        $request['status'] = filter_var($request['status'], FILTER_VALIDATE_BOOLEAN);
+
+
+        $rules = [  
+            'title' => 'string|nullable',
+            'description' => 'string|nullable',
+            'due_date' => 'date|nullable',
+            'status' => 'boolean|nullable'
+        ];
+
+        $messages = [
+            'required' => ":attribute is required. Please choose it"
+        ];
+
+        $validator = Validator::make($request->all(), $rules, $messages);
+
+        if ($validator->fails()) {
+            return response()->json(['errors' => $validator->errors(), 404]);
+        }
+        
+
         $title = $request->input("title");
         $description = $request->input("description");
         $dueDate = $request->input("due_date");
@@ -144,10 +200,10 @@ class IssueController extends Controller
         }
 
         if ($status != null) {
-            if ($status == "true") {
+            if ($status == true) {
                 $issue->is_completed = true;
                 $issue->closed_date = Carbon::now()->toDateTimeString();
-            } else if ($status == "false") {
+            } else if ($status == false) {
                 $issue->is_completed = false;
             }
         }
@@ -160,6 +216,24 @@ class IssueController extends Controller
     public function assign(Request $request, $id)
     {
         $issue = Issue::findOrFail($id);
+
+        $rules = [  
+            'user' => 'required|integer|exists:user,id',
+            'sender' => 'required|integer|exists:user,id',
+            'due_date' => 'date|nullable',
+            'status' => 'boolean|nullable'
+        ];
+
+        $messages = [
+            'required' => ":attribute is required. Please choose it"
+        ];
+
+        $validator = Validator::make($request->all(), $rules, $messages);
+
+        if ($validator->fails()) {
+            return response()->json(['errors' => $validator->errors(), 404]);
+        }
+        
 
         $user_id = $request->input("user");
         $user = User::findOrFail($user_id);
@@ -199,6 +273,20 @@ class IssueController extends Controller
     public function desassign(Request $request, $id)
     {
         $issue = Issue::findOrFail($id);
+
+        $rules = [  
+            'user' => 'required|integer|exists:user,id',
+        ];
+
+        $messages = [
+            'required' => ":attribute is required. Please choose it"
+        ];
+
+        $validator = Validator::make($request->all(), $rules, $messages);
+        if ($validator->fails()) {
+            return response()->json(['errors' => $validator->errors(), 404]);
+        }
+
         $user = $request->input("user");
         $assigned_user = AssignedUser::find(['user_id' => intval($user), 'issue_id' => intval($id)]);
         $assigned_user->delete();
@@ -209,6 +297,21 @@ class IssueController extends Controller
     public function label(Request $request, $id)
     {
         $issue = Issue::findOrFail($id);
+
+        $rules = [  
+            'label' => 'nullable|integer|exists:tag,id',
+            'name' => 'required|string|min:1'
+        ];
+
+        
+
+        $validator = Validator::make($request->all(), $rules);
+        if ($validator->fails()) {
+            return response()->json(['errors' => $validator->errors(), 404]);
+        }
+        return response()->json([$request->all()]);
+
+        
         $label = $request->input("label");
         $name = $request->input("name");
         $tag = null;
